+++ pkg/http_router/routes/auth/auth.go
@@ -66,7 +66,7 @@ func (a *AuthRoutes) login(w http.ResponseWriter, r *http.Request) {
 		}
 
 		// verify validity
-		if actualToken.Disabled {
+		if !actualToken.Enabled {
 			handlerUtils.PostErrorResponse(w, "invalid token", http.StatusUnauthorized)
 			return
 		}
@@ -107,7 +107,7 @@ func (a *AuthRoutes) login(w http.ResponseWriter, r *http.Request) {
 		userInDB = _userInDB
 	}
 
-	if userInDB.Disabled {
+	if !userInDB.Enabled {
 		handlerUtils.PostErrorResponse(w, "user is disabled", http.StatusUnauthorized)
 		return
 	}