+++ pkg/http_router/routes/auth/oauth.go
@@ -86,7 +86,7 @@ func (oa *OAuthRoutes) login(w http.ResponseWriter, r *http.Request) {
 		}
 
 		// verify validity
-		if svcAccount.Disabled {
+		if !svcAccount.Enabled {
 			handlerUtils.PostErrorResponse(w, "invalid token", http.StatusUnauthorized)
 			return
 		}
@@ -119,7 +119,7 @@ func (oa *OAuthRoutes) login(w http.ResponseWriter, r *http.Request) {
 		userInDB = _userInDB
 	}
 
-	if userInDB.Disabled {
+	if !userInDB.Enabled {
 		handlerUtils.PostErrorResponse(w, "user is disabled", http.StatusUnauthorized)
 		return
 	}