+++ pkg/upgrade/versions.go
@@ -18,4 +18,5 @@ var upgrades = map[string]upgradeFunction{
 	"2.1.1-1": upgrade_2_1_1__1, // 2.1.1 upgrade #2
 	"2.2.0-1": upgrade_2_2_0__1, // 2.2.0: RBAC policies + admin for existing users
 	"2.2.0-2": upgrade_2_2_0__2, // 2.2.0: rename service_token entity to service_account
+	"2.3.0-1": upgrade_2_3_0__1, // 2.3.0: user/service-account disabled → enabled
 }